+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * KMS Utils dependency remove from Stratio Spark Dispatcher and History Server
 * Possibility of specifying the name of the certificate within a directory with multiple certificates
 * Add support to multiples CAs
+* Added performance tests PNF
 
 ## 2.2.0.5 (upcoming)